Save played questions
This is in addition to the previous commit, which saves questions
from the system context only when they are added to a quiz.
With the previous change, if a quiz has random questions, they are
missing from the backup. With this change, we also include all played
questions in order to include random questions in the backup.